+++ pygtext.py
@@ -6,19 +6,14 @@ import pygame, string
 #pygfile for pygame printing
 #source: https://www.cs.unc.edu/~gb/blog/2007/11/16/python-file-like-object-for-use-with-print-in-pygame/
 class Pygfile(object):
-  def __init__(self, font=None, prompt='> ', cursor='_', maxlength=-1, prompt_enable=True):
+  def __init__(self, font=None, cursor='_', maxlength=-1, prompt_enable=True, parent=None):
     if font is None:
       font = pygame.font.SysFont('monospace', 10)
     self.font = font
     self.buff = []
-    self.value = ''
-    self.shift = False
-    self.ctrl = False
     #self.restricted = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789!"#$%&\\\'()*+,-./:;<=>?@[\]^_`{|}~'
-    self.maxlength = maxlength
     self.prompt_enable = prompt_enable
-    self.prompt = prompt
-    self.cursor = cursor
+    self.parent = parent
     
     inkey   = '123457890-=/;\'[]\\'
     shifted = '!@#$%&*()_+?:\"{}|'
@@ -58,7 +53,7 @@ class Pygfile(object):
     w,h = surf.get_size()
     # join all the writes together into one string
     text = ''.join(self.buff)
-    if self.prompt_enable: text = text + self.prompt + self.value + self.cursor
+    if self.prompt_enable: text = text + self.parent.terminal.makeprompt(self.parent.cursor_state)
     # and split it into lines
     lines = text.split('\n')
     # bust up any long lines into pieces that fit
@@ -82,26 +77,6 @@ class Pygfile(object):
   def clear(self):
     '''Clear the buffer (and thus the display)'''
     self.buff = []
-        
-  def updateinput(self, events):
-    '''Update the input based on passed events'''
-    for event in events:
-      if event.type == KEYUP:
-        if event.key == K_LSHIFT or event.key == K_RSHIFT: self.shift = False
-        if event.key == K_LCTRL or event.key == K_RCTRL: self.ctrl = False
-      if event.type == KEYDOWN:
-        if event.key == K_BACKSPACE: self.value = self.value[:-1]
-        elif event.key == K_LSHIFT or event.key == K_RSHIFT: self.shift = True
-        elif event.key == K_LCTRL or event.key == K_RCTRL: self.ctrl = True
-        if self.ctrl:
-          # ctrl-c clears input line
-          if event.key == K_c: self.value = ''
-        elif not self.shift:
-          if event.key in range(32,126): self.value += chr(event.key)
-        elif self.shift:
-          if event.key in range(32,126): self.value += shifted(chr(event.key), self.table)
-
-    if len(self.value) > self.maxlength and self.maxlength >= 0: self.value = self.value[:-1]
 
 def shifted(char, table):
     return char.translate(table).upper()
